requirements.txt
openai>=0.27.0
pysolr>=3.9.0
python-dotenv>=0.21.0